% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_pipeline_utils.R
\name{append_lags}
\alias{append_lags}
\title{Append n lags of a group of variables to a dataset}
\usage{
append_lags(data, n_lags, lag_vars, default = 0)
}
\arguments{
\item{data}{data frame}

\item{n_lags}{int. number of lags to append}

\item{lag_vars}{character vector. Names of dataframe columns for which you want lags}

\item{default}{vector of length 1 or nrow(data), as in the \code{default} argument in \code{dplyr::lag}.}
}
\value{
The original data frame with additional columns \code{x_lag1},\code{x_lag2},...,\verb{x_lag\{n_lags\}} for \code{x} in \code{lag_vars}.
}
\description{
Append n lags of a group of variables to a dataset
}
