% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap_se}
\alias{bootstrap_se}
\title{Bootstrap standard error estimates for didgformula estimators}
\usage{
bootstrap_se(data, nboots, estimator, ..., tibble = TRUE)
}
\arguments{
\item{data}{Wide format data frame with one row per individual, and columns Yt for t = 0,1,...,Tt.}

\item{nboots}{integer. Number of boostrap iterations}

\item{estimator}{function. One of \verb{*_pipeline()} functions.}

\item{...}{Arguments passed to \code{estimator}}

\item{tibble}{logical. Return results as a tibble (TRUE) or vector (FALSE)?}
}
\description{
Bootstrap standard error estimates for didgformula estimators
}
