% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iptw.R
\name{calc_weights}
\alias{calc_weights}
\title{Calculate inverse probability of treatment weights}
\usage{
calc_weights(denominator, numerator = 1)
}
\arguments{
\item{denominator}{Matrix of predictions, as returned by \code{pred_treatment_models()}}

\item{numerator}{Matrix of predictions, as returned by \code{pred_treatment_models()}}
}
\description{
Calculate inverse probability of treatment weights
}
