% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iptw.R
\name{calc_ydiffs}
\alias{calc_ydiffs}
\title{Calculate Y_t - Y_t-1}
\usage{
calc_ydiffs(data, Tt)
}
\arguments{
\item{data}{Wide format data frame with one row per individual, and columns Yt for t = 0,1,...,Tt.}

\item{Tt}{int. Final period in dataset (t=0,1,...,Tt)}
}
\value{
matrix with nrow(data) rows and Tt columns, each with values of Y_t - Y_t-1 for t=1,2,...,Tt.
}
\description{
Calculate Y_t - Y_t-1
}
