% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{combine_estimates}
\alias{combine_estimates}
\title{Combine estimates from different pipelines for comparison}
\usage{
combine_estimates(truth, iptw, ice, or, Tt)
}
\arguments{
\item{truth}{tibble. Return value of \code{estimate_truth()}}

\item{iptw}{tibble. Return value of \code{iptw_pipeline()}}

\item{ice}{tibble. Return value of \code{ice_pipeline()}}

\item{or}{tibble. Return value of \code{or_pipeline()}}

\item{Tt}{max periods.}
}
\value{
tibble.
}
\description{
Combine estimates from different pipelines for comparison
}
