% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_tools.R
\name{dens}
\alias{dens}
\title{Evaluate data against the implied density from a fitted \code{glm} object based on possibly new covariate values, with automatic detection of family and link.}
\usage{
dens(x, model, newdata, binomial_n = 1)
}
\arguments{
\item{x}{vector of values at which to evaluate the density}

\item{model}{a fitted \code{glm} object}

\item{newdata}{data frame with values of the covariates based on which to simulate the response, similar to \code{predict.glm}. nrow(newdata) must equal length(x).}

\item{binomial_n}{int length nrow(newdata). If \code{model} was fit with \code{link=binomial}, you can optionally pass a vector of group sizes to simulate aggregate binomial data.}
}
\value{
A vector of length(x) of density estimates.
}
\description{
Evaluate data against the implied density from a fitted \code{glm} object based on possibly new covariate values, with automatic detection of family and link.
}
\examples{
n  = 100
df = tibble::tibble(x = rnorm(n), y = rpois(n, lambda = exp(x)))
mod = glm(y~x, poisson('log'), df)
dens_y = dens(df$y, mod, newdata=df)

# equivalent to
dpois(df$y, lambda = predict(mod, type='response'))

}
