% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{estimate_bias_variance}
\alias{estimate_bias_variance}
\title{Estimate the bias and variance of DID g-formula estimators}
\usage{
estimate_bias_variance(results_df, estimates_truth)
}
\arguments{
\item{results_df}{Data frame with columns 'rep' and 'estimates' where 'estimates' is a list of tibble output results from either iptw_pipeline(), ice_pipeline(), or or_pipeline().}

\item{estimates_truth}{Data frame with columns 't' and 'estimate', with the 'true' value of E(Y_t(a)-Y_t-1(a)) for t=1,...,T}
}
\value{
data frame with bias and variance estimates
}
\description{
Estimate the bias and variance of DID g-formula estimators
}
