% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{estimate_ci_coverage}
\alias{estimate_ci_coverage}
\title{Estimate confidence interval coverage for a normality-assuming interval based on the simulation variance.}
\usage{
estimate_ci_coverage(results_df, estimates_truth, conf_level = 0.95)
}
\arguments{
\item{results_df}{Data frame with columns 'rep' and 'estimates' where 'estimates' is a list of tibble output results from either iptw_pipeline(), ice_pipeline(), or or_pipeline().}

\item{estimates_truth}{Data frame with columns 't' and 'estimate', with the 'true' value of E(Y_t(a)-Y_t-1(a)) for t=1,...,T}

\item{conf_level}{num. level of confidence, e.g. .95 is a 95\% confidence interval}
}
\value{
data frame with coverage estimates
}
\description{
Estimate confidence interval coverage for a normality-assuming interval based on the simulation variance.
}
