% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{estimate_truth}
\alias{estimate_truth}
\title{Estimate true values of the parameters from the simulated potential outcomes}
\usage{
estimate_truth(df_po, Tt, link_fun = NULL, binomial_n = 1)
}
\arguments{
\item{df_po}{data frame. Return value of \code{generate_data(potential_outcomes=TRUE)}}

\item{Tt}{max periods.}
}
\value{
tibble.
}
\description{
Estimate true values of the parameters from the simulated potential outcomes
}
