% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ice.R
\name{fit_outer_exp_model}
\alias{fit_outer_exp_model}
\title{Fit outer moderls for ICE using \code{lm}}
\usage{
fit_outer_exp_model(k, data, preds, rhs_formula, binomial_n = NULL)
}
\arguments{
\item{k}{int. Timepoint to which the outer expectation, i.e. E(mu_k^t|L_k, A_k), applies.}

\item{data}{wide data frame}

\item{preds}{vector of length(nrow(data)) of values predicted by the k+1 model.}

\item{binomial_n}{vector of group sizes if Y refers to binomial aggregate data. Used to construct weights.}

\item{rhs_formula_t}{chr. RHS glue-style formula for Yt model}

\item{rhs_formula_tmin}{chr. RHS glue-style formula for Yt-1 model}
}
\description{
Fit outer moderls for ICE using \code{lm}
}
