% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iptw.R
\name{fit_treatment_models}
\alias{fit_treatment_models}
\title{Fit treatment models for IPTW}
\usage{
fit_treatment_models(data, rhs_formula, Tt)
}
\arguments{
\item{data}{Wide format data frame with one row per individual, and columns Lt, At, Yt for t = 0,1,...,Tt.}

\item{rhs_formula}{chr. Formulas to use Format '~L{t}+L{t-1}' etc.}

\item{Tt}{int. Final period in dataset (t=0,1,...,Tt)}
}
\value{
list of models for period t=0,1,...,Tt, in that order.
}
\description{
Fit treatment models for IPTW
}
