% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generate simulated data that meets the parallel trends assumption}
\usage{
generate_data(
  N,
  Tt,
  Beta,
  potential_outcomes = FALSE,
  ylink = "rnorm_identity",
  binomial_n = 1,
  long = FALSE
)
}
\arguments{
\item{N}{int. Number of independent observations}

\item{Tt}{int. Number of periods, minus 1. I.e. there are Tt + 1 periods.}

\item{Beta}{list of length 4. Output of generate_parameters().}

\item{potential_outcomes}{logical. Should outcomes and covariates be generated with exposure set to 0 at all times?}

\item{ylink}{chr. One of "rnorm_identity", "rbinom_logit", or "rbinom_logit_hazard".}

\item{binomial_n}{int length N. Defaults to all 1's. If ylink is rbinom_logit, you can optionally pass a vector of group sizes to generate aggregate binomial data. In this case, treatments and covariates will be constant at the group level for a given time period.}

\item{long}{lgl. Should the returned dataset be wide (one row per participant, FALSE), or long (Tt+1 rows per participant, TRUE) ?}
}
\value{
Data frame with N rows and (Tt+1)*3 + 2 columns - 'uid' is a unique identifier, 'U0' is an 'unmeasured' baseline covariate, L{t},A{t},Y{t} are covariates, exposures, and outcomes, respectively. If binomial_n != 1, an additional column binomial_n is also included.
}
\description{
Generate simulated data that meets the parallel trends assumption
}
