% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_parameters.R
\name{generate_parameters}
\alias{generate_parameters}
\title{Pick parameter values for simulation, from a normal distribution}
\usage{
generate_parameters(
  Tt,
  mu_Beta_L = 0.2,
  mu_Beta_W = 0.2,
  mu_Beta_A = 0.2,
  mu_Beta_Y = 0.2,
  sd_Beta_L = 0.2,
  sd_Beta_W = 0.2,
  sd_Beta_A = 0.2,
  sd_Beta_Y = 0.2,
  range_ymeans = c(-5, 5),
  constant_dydu = 0.1,
  check_CDF = FALSE,
  check_rbinom_identity = FALSE
)
}
\arguments{
\item{Tt}{Final period (t=0,1,...,Tt)}

\item{mu_Beta_L}{numeric. mean for parameters in covariate models}

\item{mu_Beta_A}{numeric. mean for parameters in treatment models}

\item{mu_Beta_Y}{numeric. mean for parameters in outcome models}

\item{sd_Beta_L}{numeric. standard deviation for parameters in covariate models}

\item{sd_Beta_A}{numeric. standard deviation for parameters in treatment models}

\item{sd_Beta_Y}{numeric. standard deviation for parameters in outcome models}

\item{ylink}{}
}
\value{

}
\description{
Pick parameter values for simulation, from a normal distribution
}
