% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_tools.R
\name{get_densfun}
\alias{get_densfun}
\title{Generate a function that looks up a density in accordance with the specification in a \code{stats::family} object}
\usage{
get_densfun(family)
}
\arguments{
\item{family}{an object of type \code{stats::family}}
}
\value{
A function with arguments \code{x} (data), \code{eta}, (linear predictor from \code{glm} object), \code{dispersion} (\code{summary(m)$dispersion} where \code{m} is a fitted \code{glm}), and  \code{n}, vector of group sizes (optional unless family=binomial)
}
\description{
This function is useful for evaluating data against the implied density of a fitted \code{glm} object
}
\examples{

fam = Gamma('log')
densfun = get_densfun(fam)
densfun(0.2, eta = 2, dispersion=0.5)

# the above is equivalent to:
dgamma(0.2, scale = exp(2), shape = 1/0.5)
}
