% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_tools.R
\name{get_simfun}
\alias{get_simfun}
\title{Generate a function that simulates data in accordance with the specification in a \code{stats::family} object}
\usage{
get_simfun(family)
}
\arguments{
\item{family}{an object of type \code{stats::family}}
}
\value{
A function that can receive a linear predictor as input and return simulated samples based on the family and link specification in the \code{stats::family} object. The arguments of this function are: \code{eta}, vector of linear predictors; \code{dispersion}, the return value of \code{summary(model)$dispersion}, where \code{model} is a fitted \code{glm} object; and \code{n}, vector of group sizes (optional unless family=binomial)
}
\description{
This function is useful for simulating data that will follow a certain \code{glm} model
}
\examples{

fam = Gamma('log')
simfun = get_simfun(fam)
sims = simfun(eta = 1, dispersion=1)

}
