% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_tools.R
\name{glue_formula}
\alias{glue_formula}
\title{Create formulas using \code{glue} syntax}
\usage{
glue_formula(string, ...)
}
\arguments{
\item{string}{chr. Expressions enclosed in braces will be evaluated as R code (see \code{?glue}).}

\item{...}{additional arguments. Here you can pass variables to be evaluated in enclosed expressions in \code{string}.}
}
\value{

}
\description{
This function is specifically designed for dynamically evaluating formulas with time indices. As such, it will delete any formula elements referring to times less than 0.
}
\examples{

<<<<<<< HEAD
glue_formula('~L{t}*L{t-1}+L{t-2}', t=2)
glue_formula('~L{t}*L{t-1}+L{t-2}', t=1)
=======
glue_formula('~L{t}+L{t-1}+L{t-2}', t=2)
glue_formula('~L{t}+L{t-1}+L{t-2}', t=1)
>>>>>>> 69b91f462779b7dc191d55589db9f217c69a68f1

}
