% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ice_long.R
\name{ice_pipeline_long}
\alias{ice_pipeline_long}
\title{Iterated conditional DID g-formula estimator}
\usage{
ice_pipeline_long(
  df_obs,
  df_interv,
  inside_formula_t,
  inside_formula_tmin1,
  outside_formula,
  Tt,
  t_col,
  n_nested = Tt,
  inside_family = "gaussian",
  pt_link_fun = NULL,
  binomial_n = NULL,
  tibble = TRUE,
  models = TRUE,
  tmle = FALSE,
  weights = NULL
)
}
\arguments{
\item{inside_formula_t}{chr, right-hand-side formula for inside model for Yt}

\item{inside_formula_tmin1}{chr, right-hand-side formula for inside model for Yt-1}

\item{outside_formula}{chr, right-hand-side formula for outside models}

\item{Tt}{int. max periods}

\item{n_nested}{int. How many nested expectations should be estimated, starting from the innermost (=0) to the outermost (=Tt)?}

\item{inside_family}{stats::family object or string referring to one, as in \code{glm}.}

\item{pt_link_fun}{function. The scale on which parallel trends is assumed (e.g., \code{qlogis} for logit scale). Default \code{NULL} for untransformed scale.}

\item{binomial_n}{int length nrow(data). Group sizes for binomial aggregate data.}

\item{tibble}{logical. return results as a tibble (TRUE) or vector (FALSE)?}

\item{models}{lgl. Return all models as an attribute?}

\item{data}{Wide format data frame with one row per individual, and columns Yt for t = 0,1,...,Tt.}
}
\value{

}
\description{
Iterated conditional DID g-formula estimator
}
