% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iptw_long.R
\name{iptw_pipeline_long}
\alias{iptw_pipeline_long}
\title{Estimate the DID g-formula using inverse-probability-of-treatment-weights with exposure models pooled over time periods.}
\usage{
iptw_pipeline_long(
  df_obs,
  df_interv,
  den_formula,
  num_formula,
  family,
  yvar = "Y",
  ylagvar = "Y_lag",
  idvar = "uid",
  timevar = "t",
  tibble = TRUE,
  pt_link_fun = NULL,
  binomial_n = NULL,
  models = TRUE
)
}
\arguments{
\item{df_obs}{long format data frame, ordered by person id, then time}

\item{df_interv}{same data set as df_obs with exposure variable set to the intervened status}

\item{den_formula}{chr or formula. \code{glm} formula for (pooled) denominator model}

\item{num_formula}{chr or formula. \code{glm} formula for (pooled) numerator model}

\item{family}{stats::families object for numerator and denominator \code{glm} call}

\item{yvar}{chr. Name of column in \code{df_obs} corresponding to the outcome at time \code{t}}

\item{ylagvar}{chr. Name of column in \code{df_obs} corresponding to the outcome at time \code{t-1}}

\item{idvar}{chr. Name of column in \code{df_obs} corresponding to the person id}

\item{timevar}{chr. Name of column in \code{df_obs} corresponding to times (typically starting from 1, so that t-1=0)}

\item{tibble}{lgl. Return results as a tibble (TRUE) or vector (FALSE)?}

\item{pt_link_fun}{function. The scale on which parallel trends is assumed (e.g., \code{qlogis} for logit scale). Default \code{NULL} for untransformed scale.}

\item{binomial_n}{int length nrow(data). Group sizes for aggregate binomial data.}

\item{models}{logical. Return all models as an attribute?}
}
\value{
Estimates of E(Yt(a) - Yt-1(a)), in the form of a tibble or vector (dependening on argument tibble), for times t=1,2,...,Tt (in that order).
}
\description{
Estimate the DID g-formula using inverse-probability-of-treatment-weights with exposure models pooled over time periods.
}
