% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_pipeline_utils.R
\name{make_time_dummies}
\alias{make_time_dummies}
\title{Add time dummy variables to a long dataset with a numeric time variable}
\usage{
make_time_dummies(df, timevar)
}
\arguments{
\item{df}{long dataset}

\item{timevar}{chr. Name of column containing times.}
}
\value{
data frame with additional columns {timevar}{n} for all unique(n) in df[\link{timevar}].
}
\description{
Add time dummy variables to a long dataset with a numeric time variable
}
