% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_regression.R
\name{or_pipeline}
\alias{or_pipeline}
\title{Outcome regression DID g-formula estimator}
\usage{
or_pipeline(
  data,
  yt_formula,
  ytmin1_formula,
  l_formula,
  y_family = gaussian,
  l_family = binomial,
  Tt,
  nreps,
  pt_link_fun = NULL,
  binomial_n = NULL,
  tibble = TRUE
)
}
\arguments{
\item{data}{Wide format data frame with one row per individual, and columns Yt for t = 0,1,...,Tt.}

\item{yt_formula}{chr. right-hand-side formula for Yt outcome modles}

\item{ytmin1_formula}{chr. right-hand-side formula for Ytmin1 outcome models}

\item{l_formula}{chr. right hand side formula for covariate models}

\item{y_family}{\code{stats::family} object for the outcome regression models (which are fit using \code{glm})}

\item{l_family}{\code{stats::family} object for the covariate models (which are fit using \code{glm})}

\item{Tt}{int. Max periods}

\item{nreps}{int. Number of reps for monte-carlo simulation}

\item{pt_link_fun}{function. The scale on which parallel trends is assumed (e.g., \code{qlogis} for logit scale). Default \code{NULL} for untransformed scale.}

\item{tibble}{logical. Return results as tibble (TRUE) or vector (FALSE)?}
}
\description{
Outcome regression DID g-formula estimator
}
