% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iptw.R
\name{pred_treatment_models}
\alias{pred_treatment_models}
\title{Get predictions from a list of treatment models}
\usage{
pred_treatment_models(data, treatment_models)
}
\arguments{
\item{data}{A data.frame generated by \code{generate_data()}}

\item{treatment_models}{A list of treatment models generated by \code{fit_treatment_models()}}
}
\description{
Get predictions from a list of treatment models
}
