% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ice.R
\name{recursive_ice}
\alias{recursive_ice}
\title{Recursively apply the ICE algorithm}
\usage{
recursive_ice(
  t,
  k,
  data,
  inside_formula,
  outside_formula,
  inside_family,
  binomial_n = NULL
)
}
\arguments{
\item{t}{int. Timepoint the estimand applies to}

\item{k}{int. Always 0 if called at the user level. Timepoint the outer expectation applies to}

\item{data}{wide data frame}

\item{inside_formula}{chr. glue-style formula for inside models (times indexed by t, e.g. '~L{t}')}

\item{outside_formula}{chr. glue-style formula for outside models (times indexed by k, e.g. '~L{k})}

\item{inside_family}{stats::families object (or chr) for glm inside models}

\item{binomial_n}{vector of length nrow(data) of group sizes if Y is binomial aggregate data}
}
\value{

}
\description{
Recursively apply the ICE algorithm
}
