% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{results_histograms}
\alias{results_histograms}
\title{Get histograms of DID g-formula results}
\usage{
results_histograms(results_df, estimates_truth, show_bias = TRUE)
}
\arguments{
\item{results_df}{Data frame with columns 'rep' and 'estimates' where 'estimates' is a list of tibble output results from either iptw_pipeline(), ice_pipeline(), or or_pipeline().}

\item{estimates_truth}{Data frame with columns 't' and 'estimate', with the 'true' value of E(Y_t(a)-Y_t-1(a)) for t=1,...,T}

\item{show_bias}{Logical. Should the histograms show deviations from 'truth' (TRUE) or the estimates themselves (FALSE)?}
}
\description{
Red vertical lines represent the true expected value for the random variable of which the displayed quantities should be realizations. If show_bias=TRUE, this is 0, otherwise it is E(Y_t(a)-Y_t-1(a)).
Blue vertical lines represent the estimated expected value based on the realizations.
}
