% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_tools.R
\name{sim}
\alias{sim}
\title{Simulate from a fitted \code{glm} object based on new data, with automatic detection of family and link.}
\usage{
sim(model, newdata, binomial_n = 1)
}
\arguments{
\item{model}{a fitted \code{glm} object}

\item{newdata}{data frame with values of the covariates based on which to simulate the response, similar to \code{predict.glm}}

\item{binomial_n}{int length nrow(newdata). If \code{model} was fit with \code{link=binomial}, you can optionally pass a vector of group sizes to simulate aggregate binomial data.}
}
\value{
A vector of length nrow(newdata) with simulated responses.
}
\description{
Simulate from a fitted \code{glm} object based on new data, with automatic detection of family and link.
}
\examples{
n  = 100
df = tibble::tibble(x = rnorm(n), y = rpois(n, lambda = exp(x)))
mod = glm(y~x, poisson('log'), df)
sims = sim(mod, newdata=df)
qqplot(df$y, sims)

}
