% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_tools.R
\name{expand_formula}
\alias{expand_formula}
\title{Create formulas by expanding out string vector arguments}
\usage{
expand_formula(formula, ...)
}
\arguments{
\item{formula}{An R formula containing terms referencing objects defined in the parent environment or in ... Terms must each consist of a single lowercase or uppercase letter.}

\item{...}{Variables referenced in \code{formula}. Can be glue-style, with glue terms also defined in ... (see examples)}
}
\value{
chr.
}
\description{
Create formulas by expanding out string vector arguments
}
\examples{
expand_formula(~a:b + d, a=c('a1','a2'), b=c('b1','b2'), d='d1')

#we could have equivalently defined any of the variables in the formula in the calling environment:
a=c('a1','a2')
b=c('b1','b2')
d='d1'
expand_formula(~a:b + d)

#variables can be specified in glue-style
expand_formula(~a:b, a=c('a{h}', 'a{h+1}'), b='a{h}', h=1)
}
