% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generate simulated data that meets the parallel trends assumption}
\usage{
generate_data(N, Tt, Beta, potential_outcomes = FALSE)
}
\arguments{
\item{N}{int. Number of independent observations}

\item{Tt}{int. Number of periods, minus 1. I.e. there are Tt + 1 periods.}

\item{Beta}{list of length 4. Output of generate_parameters().}

\item{potential_outcomes}{logical. Should outcomes and covariates be generated with exposure set to 0 at all times?}
}
\value{
Data frame with N rows and (Tt+1)*3 + 2 columns - 'uid' is a unique identifier, 'U0' is an 'unmeasured' baseline covariate, L{t},A{t},Y{t} are covariates, exposures, and outcomes, respectively.
}
\description{
Generate simulated data that meets the parallel trends assumption
}
