% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ice.R
\name{ice_pipeline}
\alias{ice_pipeline}
\title{Iterated conditional DID g-formula estimator}
\usage{
ice_pipeline(
  data,
  inside_formula,
  outside_formula,
  Tt,
  inside_family = "gaussian",
  tibble = TRUE
)
}
\arguments{
\item{data}{Wide format data frame with one row per individual, and columns Yt for t = 0,1,...,Tt.}

\item{inside_formula}{chr, right-hand-side formula for inside models}

\item{outside_formula}{chr, right-hand-side formula for outside models}

\item{Tt}{int. max periods}

\item{tibble}{logical. return results as a tibble (TRUE) or vector (FALSE)?}
}
\value{

}
\description{
Iterated conditional DID g-formula estimator
}
