% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iptw.R
\name{iptw_pipeline}
\alias{iptw_pipeline}
\title{Estimate the DID g-formula using inverse-probability-of-treatment-weights.}
\usage{
iptw_pipeline(data, rhs_formula, Tt, tibble = TRUE)
}
\arguments{
\item{data}{Wide format data frame with one row per individual, and columns Lt, At, Yt for t = 0,1,...,Tt.}

\item{rhs_formula}{chr. Right-hand-side formula to use in predicting the treatment. Format: '~L\{t\}+L\{t-1\}' etc.}

\item{Tt}{int. Final period in dataset (t=0,1,...,Tt)}

\item{tibble}{logical. Should the results be returned as a tibble with columns (t, estimates) (TRUE) or a vector of just the estimates  (FALSE)?}
}
\value{
Estimates of E(Yt(a) - Yt-1(a)), in the form of a tibble or vector (dependening on argument tibble), for times t=1,2,...,Tt (in that order).
}
\description{
Estimate the DID g-formula using inverse-probability-of-treatment-weights.
}
