% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_regression.R
\name{or_pipeline}
\alias{or_pipeline}
\title{Outcome regression DID g-formula estimator}
\usage{
or_pipeline(
  data,
  y_formula,
  l_formula,
  y_family = "gaussian",
  l_family = "binomial",
  Tt,
  nreps,
  binomial_n = NULL,
  tibble = TRUE
)
}
\arguments{
\item{data}{Wide format data frame with one row per individual, and columns Yt for t = 0,1,...,Tt.}

\item{y_formula}{chr. right-hand-side formula for outcome models}

\item{l_formula}{chr. right hand side formula for covariate models}

\item{Tt}{int. Max periods}

\item{nreps}{int. Number of reps for monte-carlo simulation}

\item{tibble}{logical. Return results as tibble (TRUE) or vector (FALSE)?}
}
\value{

}
\description{
Outcome regression DID g-formula estimator
}
