% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ice.R
\name{recursive_ice}
\alias{recursive_ice}
\title{Recursively apply the ICE algorithm}
\usage{
recursive_ice(
  t,
  k,
  data,
  inside_formula_t,
  inside_formula_tmin1,
  outside_formula,
  inside_family,
  binomial_n = NULL,
  models = TRUE,
  tmle = FALSE,
  weights = NULL
)
}
\arguments{
\item{t}{int. Timepoint the estimand applies to}

\item{k}{int. Always 0 if called at the user level. Timepoint the outer expectation applies to}

\item{data}{wide data frame}

\item{inside_formula_t}{chr. glue-style formula for inside model for Yt (times indexed by t, e.g. '~L{t}')}

\item{inside_formula_tmin1}{chr. glue-style formula for inside model for Yt (times indexed by t, e.g. '~L{t-1}')}

\item{outside_formula}{chr. glue-style formula for outside models (times indexed by k, e.g. '~L{k})}

\item{inside_family}{stats::families object (or chr) for glm inside models}

\item{binomial_n}{vector of length nrow(data) of group sizes if Y is binomial aggregate data}

\item{models}{lgl. Return all models as an attribute?}

\item{tmle}{lgl. Incorporate tmle updating step to make estimator doubly robust?}

\item{weights}{Matrix of IPTW weights as returned by \code{cbind(1, calc_weights(...))}. Default \code{NULL}, required if \code{tmle=TRUE}.}
}
\value{

}
\description{
Recursively apply the ICE algorithm
}
